refactor(hooks): revert file name useHistoryManagerWithActivityMonitoring to useHistoryManagerWithActivity
