+++ packages/cli/src/ui/hooks/useHistoryManagerWithActivity.ts
